RavenDB-17674 RavenDB and Prometheus integration - moved shared logic to Metrics Provider
